Minor code reformatting / moving
